Dry run mode on fs module

+++ m2bk/fs.py
@@ -26,7 +26,10 @@ from .const import PKG_NAME
 FS_DEFAULT_OUTPUT_DIR = "/tmp/{pkg_name}".format(pkg_name=PKG_NAME)
 
 # Output directory name
-_output_dir = None
+_output_dir = FS_DEFAULT_OUTPUT_DIR
+
+# dry run mode
+_dry_run = False
 
 
 def get_output_dir():
@@ -38,7 +41,7 @@ def get_output_dir():
     return _output_dir
 
 
-def init(*, output_dir=FS_DEFAULT_OUTPUT_DIR, **kwargs):
+def init(*, output_dir=FS_DEFAULT_OUTPUT_DIR, dry_run=False, **kwargs):
     """
     Set up output directory
 
@@ -48,6 +51,10 @@ def init(*, output_dir=FS_DEFAULT_OUTPUT_DIR, **kwargs):
     # Output directory
     global _output_dir
     _output_dir = output_dir
+
+    # Dry run mode
+    global _dry_run
+    _dry_run = dry_run
 
     # Type checks
     utils.chkstr(_output_dir, 'output_dir')
@@ -60,10 +67,11 @@ def init(*, output_dir=FS_DEFAULT_OUTPUT_DIR, **kwargs):
     if not os.path.exists(_output_dir):
         log.msg_warn("Output path '{out_dir}' does not exist, creating it ..."
                      .format(out_dir=_output_dir))
-        # create the actual root output directory
-        os.makedirs(_output_dir)
-        # set folder permissions to 0770
-        os.chmod(_output_dir, stat.S_IRWXU | stat.S_IRWXG)
+        if not _dry_run:
+            # create the actual root output directory
+            os.makedirs(_output_dir)
+            # set folder permissions to 0770
+            os.chmod(_output_dir, stat.S_IRWXU | stat.S_IRWXG)
 
 
 def cleanup():
@@ -71,7 +79,8 @@ def cleanup():
     if _output_dir and os.path.exists(_output_dir):
         log.msg_warn("Cleaning up output directory at '{output_dir}' ..."
                      .format(output_dir=_output_dir))
-        shutil.rmtree(_output_dir)
+        if not _dry_run:
+            shutil.rmtree(_output_dir)
 
 
 def make_tmp_dir(prefix):
@@ -94,7 +103,8 @@ def make_tmp_dir(prefix):
                                               now=now, uuid=f_uuid)
 
     # create the actual directory
-    os.makedirs(out_dir)
+    if not _dry_run:
+        os.makedirs(out_dir)
     log.msg_debug("Output directory: {out_dir}".format(out_dir=_output_dir))
 
     # Return a string containing the name and path of the output directory
@@ -112,6 +122,7 @@ def make_tarball(src_dir):
         raise TypeError('src_dir must be str')
     output_file = src_dir + ".tar.gz"
     log.msg("Wrapping tarball '{out}' ...".format(out=output_file))
-    with tarfile.open(output_file, "w:gz") as tar:
-        tar.add(src_dir, arcname=os.path.basename(src_dir))
+    if not _dry_run:
+        with tarfile.open(output_file, "w:gz") as tar:
+            tar.add(src_dir, arcname=os.path.basename(src_dir))
     return output_file
